implement custom useSafeAreaInsets and adapt SafeAreaConsumer on Android
