+++ src/hooks/useSafeAreaInsets/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable no-restricted-imports */
+import {useSafeAreaInsets} from 'react-native-safe-area-context';
+
+export default useSafeAreaInsets;
